requirements.txt
# Web框架
fastapi==0.68.0
uvicorn==0.15.0

# 网页抓取和解析
beautifulsoup4==4.9.3
requests==2.25.1
selenium==3.141.0

# PDF处理
PyPDF2==1.26.0
pdfplumber==0.5.28
python-docx==0.8.11
reportlab==3.5.68

# 自然语言处理
nltk==3.6.2
jieba==0.42.1

# 数据处理
pandas==1.3.0
numpy==1.21.0

# 异步任务处理
celery==5.1.2

# 数据库
sqlalchemy==1.4.22
psycopg2==2.9.1

# 工具类
python-dotenv==0.18.0